+++ nens_auth_client/cognito.py
@@ -148,3 +148,24 @@ class CognitoOAuthClient(DjangoOAuth2App):
 
         claims.validate(leeway=leeway)
         return claims
+
+    @staticmethod
+    def extract_provider_name(claims):
+        """Return provider name from claim and `None` if not found"""
+        # Also used by backends.py
+        try:
+            return claims["identities"][0]["providerName"]
+        except (KeyError, IndexError):
+            return
+
+    @staticmethod
+    def extract_username(claims) -> str:
+        """Return username from claims"""
+        username = ""
+        if claims.get("identities"):
+            # External identity providers result in usernames that are not
+            # recognizable by the end user. Use the email instead.
+            username = claims.get("email")
+        if not username:
+            username = claims["cognito:username"]
+        return username