+++ nens_auth_client/testsettings.py
@@ -118,11 +118,10 @@ INSTALLED_APPS = (
 )
 
 # Add your production name here
-ALLOWED_HOSTS = ["localhost"]
+ALLOWED_HOSTS = ["localhost", "0.0.0.0"]
 
 AUTHENTICATION_BACKENDS = [
     "nens_auth_client.backends.RemoteUserBackend",
-    "nens_auth_client.backends.SSOMigrationBackend",
     "django.contrib.auth.backends.ModelBackend",
 ]